Remove HWP Rotation Agent
Also split the old hwp rotation agent tests into separate tests for the PID and
PMX agents.
