+++ docs/index.rst
@@ -43,7 +43,6 @@ API Reference        Full API documentation for core parts of the SOCS library.
     agents/fts_agent
     agents/hwp_encoder
     agents/hwp_picoscope
-    agents/hwp_rotation_agent
     agents/hwp_pid
     agents/hwp_pmx
     agents/holo_fpga
