+++ tests/default.yaml
@@ -36,13 +36,18 @@ hosts:
        'arguments': [['--ip-address', '127.0.0.1'],
                        ['--port-number', '9001'],
                        ['--turbo-address', '1']]},
-      {'agent-class': 'RotationAgent',
-       'instance-id': 'rotator',
-       'arguments': [['--kikusui-ip', '127.0.0.1'],
-                     ['--kikusui-port', '2000'],
-                     ['--pid-ip', '127.0.0.1'],
-                     ['--pid-port', '2001'],
+      {'agent-class': 'HWPPIDAgent',
+       'instance-id': 'hwp-pid',
+       'arguments': [['--ip', '127.0.0.1'],
+                     ['--port', '2001'],
                      ['-v'],
+                     ['--mode', 'init'],
+                    ]
+      },
+      {'agent-class': 'HWPPMXAgent',
+       'instance-id': 'hwp-pmx',
+       'arguments': [['--ip', '127.0.0.1'],
+                     ['--port', '2000'],
                      ['--mode', 'init'],
                     ]
       },
